+++ GPOTools/changelog.md
@@ -1,3 +1,10 @@
 ﻿# Changelog
+## 0.3.0 (2020-07-05)
+
+- Upd: Network Share remapping
+- Fix: Identity resolution - Now works in Subdomains
+- Fix: Identity resolution - various issues
+
 ## 0.1.0 (2019-06-29)
- - Initial release
+
+- Initial release